Fix Year 2038 problem for future installs (#2126)

+++ common/src/main/resources/me/lucko/luckperms/schema/h2.sql
@@ -7,7 +7,7 @@ CREATE TABLE `{prefix}user_permissions` (
   `value`      BOOL               NOT NULL,
   `server`     VARCHAR(36)        NOT NULL,
   `world`      VARCHAR(36)        NOT NULL,
-  `expiry`     INT(11)            NOT NULL,
+  `expiry`     BIGINT             NOT NULL,
   `contexts`   VARCHAR(200)       NOT NULL,
   PRIMARY KEY (`id`)
 );
@@ -20,7 +20,7 @@ CREATE TABLE `{prefix}group_permissions` (
   `value`      BOOL               NOT NULL,
   `server`     VARCHAR(36)        NOT NULL,
   `world`      VARCHAR(36)        NOT NULL,
-  `expiry`     INT(11)            NOT NULL,
+  `expiry`     BIGINT             NOT NULL,
   `contexts`   VARCHAR(200)       NOT NULL,
   PRIMARY KEY (`id`)
 );
